use sidekiq testing framework to run jobs inline in development and fake Redis in test (#636)
* Fake Sidekiq redis in test

This adds `require Sidekiq/Testing` to the spec helper. This automatically calls `Sidekiq::Testing.fake!` and makes it so Sidekiq doesn't try to hit Redis in tests. More on Sidekiq test in https://github.com/mperham/sidekiq/wiki/Testing .

* Run Sidekiq jobs inline in development

This Sidekiq config runs our jobs inline in development. Postmark prints emails to the log.

* Stop checking if ENV["REDIS_URL"] is set

Since we have Sidekiq configured to work in development and test where we don't have Redis setup, we don't need to check this.

* Stop manipulating ENV["REDIS_URL"] in tests

Since we no longer check if ENV["REDIS_URL"] is set, we don't need to set it in tests.

* Ignore config and spec files when computing codecov coverage.

* update README to describe Sidekiq

Co-authored-by: Ana Ulin <ana@ulin.org>