(694) fix Sidekiq Logging and Stripe Reconciliation  (#643)
* log to Sidekiq logger in Sidekiq jobs

This sets Rails.logger to return the Sidekiq logger when in a Sidekiq process. Previously, it would default to the normal Rails logger and attempts to log disappeared into the void.

* Add Sidekiq instructions to README

* Use OpenStruct for Stripe objects in reconciliation

We were using symbols when we should have been using strings during Stripe reconciliation. Let's just use OpenStruct instead and avoid the whole issue.

* Add more logging to Stripe reconciliation